norminetizando

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: anacaro3 <anacaro3@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/07 00:51:36 by coder             #+#    #+#             */
-/*   Updated: 2022/09/14 00:59:32 by anacaro3         ###   ########.fr       */
+/*   Updated: 2022/09/15 00:48:25 by anacaro3         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char * ft_strnstr(const char *big, const char *little, size_t len)
 {
-    int i;
-    int j;
+    unsigned long int i;
+    unsigned long int j;
 
     i = 0;
     j = 0;
